Bilhetagem - Adiciona Transações do RioCard na `passageiros_hora` (#130)
* atualizacao horaria transacao riocard

* add riocard nas tabelas de passageiros_hora

* corrige transacao_riocard

* otimiza join geografico

* remove view_passageiros_hora_riocard e retira filtro do riocard de view_passageiros_hora

* retira riocard da view do dash

* add modo no schema

* muda aux sufixo para prefixo

* remove view deletada

* altera flow materializacao validacao riocard

* add changelog

* altera data dbt

* remove transacoes riocard

* altera source para prod

---------

+++ queries/models/br_rj_riodejaneiro_bilhetagem_staging/staging_transacao_riocard.sql
@@ -34,9 +34,9 @@ SELECT
     SAFE_CAST(JSON_VALUE(content, '$.tipo_integracao') AS STRING) AS tipo_integracao,
     SAFE_CAST(JSON_VALUE(content, '$.tipo_transacao') AS STRING) AS tipo_transacao,
     SAFE_CAST(JSON_VALUE(content, '$.uid_origem') AS STRING) AS uid_origem,
-    SAFE_CAST(JSON_VALUE(content, '$.valor_tarifa') AS FLOAT64) AS valor_tarifa,
-    SAFE_CAST(JSON_VALUE(content, '$.valor_transacao') AS FLOAT64) AS valor_transacao,
+    SAFE_CAST(JSON_VALUE(content, '$.valor_tarifa') AS NUMERIC) AS valor_tarifa,
+    SAFE_CAST(JSON_VALUE(content, '$.valor_transacao') AS NUMERIC) AS valor_transacao,
     SAFE_CAST(JSON_VALUE(content, '$.veiculo_id') AS STRING) AS veiculo_id,
-    SAFE_CAST(JSON_VALUE(content, '$.vl_saldo') AS FLOAT64) AS vl_saldo
+    SAFE_CAST(JSON_VALUE(content, '$.vl_saldo') AS NUMERIC) AS vl_saldo
 FROM
     {{ source("br_rj_riodejaneiro_bilhetagem_staging", "transacao_riocard") }}
